requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-import-export==2.0.2
django-restframework==0.0.1
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
idna==2.7
oauthlib==2.1.0
odfpy==1.4.1
openpyxl==3.0.3
pytz==2019.3
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.0.0
sqlparse==0.3.1
urllib3==1.24.2
tablib==1.1.0
xlrd==1.2.0
xlwt==1.3.0
djangorestframework==3.11.0
pillow

whitenoise
six
s3transfer
gunicorn==20.0.4
django-heroku==0.3.1
boto3
django-storages==1.8
Jinja2==2.10.3

django-cors-headers
django-rest-auth
django-allauth

django-nested-admin
razorpay
Celery
redis==2.10.3
django-extra-fields




tensorflow==2.2.0
keras==2.4.3
opencv-python==4.5.1.48
nltk
SpeechRecognition
pyaudio
numpy
scikit-learn==0.24.1
wget
protobuf==3.8.0
